+++ flask/users.py
@@ -58,5 +58,10 @@ def update_user(id):
 @users_bp.route('/users/<id>', methods=['DELETE'])
 def delete_user(id):
     collection = db.users
+    user = collection.find_one({'_id': ObjectId(id)})
+    if user is None:
+        return jsonify({"message": "User not found"}), 404
     result = collection.delete_one({'_id': ObjectId(id)})
-    return jsonify(result)
+    if result.deleted_count == 0:
+        return jsonify({"message": "User deleted failed"}), 500
+    return jsonify({"message": "User deleted successfully"}), 200
